Added source code for EntityGraph
The entity graph concept enables defining graphs of related entities which form a unit. Entity graph methods operate an entity graph as a whole. For example, detaching an entity from a context, will detach all entities in that graph from the context. Cloning an entity will clone all entities in the graph, and deleting an entity will delete all entities in the graph (similar to 'cascading deletes' in data bases).