EntityGraph: Workaround for RIA bug resulted in attaching a clone to the domain context
RIA doesn't correctly apply relationship span to newly created entities. This may result in entity graphs where entities have a non-null foreign key, but the corresponding navigation property is null.
    
The workaround, that explicitly sets the navigation property, didn't check if the target entity is attached to a context. If it is, assigning the navigation property automatically attaches the complete entity graph to the domain context. Which is not the intended behavior.
    
Fix. Only set the navigation property if the target entity is detached.