Renamed Clone method to Copy.
Since the clone method does not copy Key properties, the result was
actually a copy of the original graph. That is, the result was an entity
graph with new entities (with new keys) with the same structure and values
as the source entity graph. Actually, this is copying, not cloning. Hence
the renaming.