EntityGraph: Reimplemtation of entity graph validator
This new validator is much more powerful the the previous one and
also solves some of its shortcomings. Most noticeable
is that the set of input expressions of a validation rule no longer
start with the root entity of an entity graph. The input expressions
now form a forest and can start from any entity that is contained in
an entity graph.
    
The following convention is used:
- If two expressions have the same parameter name, they start from the same object,
For example:
     A => A.<path1>
     A => A.<path2?
Are two dependency expressions. Since they have the same parameter name ('A') they will
start from the same object.
    
     A => A.<path1>
     B => B.<path2>
In this case both expressions are independent.