EntityGraph: Changed names spaces and removed TEntity type parameter
Non-RIA specific stuff is now in namespace RiaServicesContrib,
RIA-specific stuff is in RiaServicesContrib.DomainServices.Client.

EntityGraphs used to be type-specific for the source entity. E.g.,
A a;
a.EntityGraph() was of type EntityGraph<A>.
This has now changed to be type EntityGraph.

For RiaServicesContrib.EntityGraph this means that it has changed from 
RiaServicesContrib.EntityGraph<TEntity,TBase,TValidationResult> to
RiaServicesContrib.EntityGraph<TEntity,TValidationResult>, where
Entity now  plays the role of TBase before.