EntityGraph: Factored out data validation from entity graph
Created a new project from the data validation stuff that was
before included in the source tree of EntityGraph.

Since the data validation framework is of more general use than for
EntityGraph only, it makes sense to put it in a separate DataValidation
project. 

EntityGraph now has a dependency on this project to be able to 
instantiate data validation for entity graphs