EntityGraph: Necessary changes to data validation framework before adding support for asynchronous data validation.
The validation framework was a mixture of asynchronous and synchronous mechanisms:
On the one hand, were validation errors reported to the framework using events, on the
other hand did each validation rule hold state information that was used during validation rule evaluation.
As a consequence the implementation was more complex than needed and asynchronous validation did not work.

In this commit I've simplified the framework by removing all asynchronous mechanisms. This version
of the framework therefore only supports synchronous validation. In a subsequent commit I will add support for
asynchronous validation by introducing a special subclass of ValidationRule.