Created a unit test framework for RiaServicesContrib
This framework is basically an empty Silverlight test project with a customized initialization method in App.xaml.cs/Application_Startup. In this method the UnitTestSystem is initialized to load all assemblies in the XAP that contain test classes (classes which are decorated with the TestClassAttribute). This means that to add new tests to the RiaServicesContrib unit tests you only need to add a reference to the project/assembly containing your tests. 
Initially I've added references to EntityTools, DataValidationFrameWork, and EntityGraph