EntityGraph: Bug fix
Trying to add a cloned entity tree that has Entities with 
"ConcurrencyMode = fixed" properties will cause a validation exception to 
be thrown in the EntityGraph.ApplyState() method. Validation needs to be 
disabled in the same way as it is done in the RiaServicesExtensions.ApplyState() 
method (i.e. by using the OnDeserializing/OnDeserialized methods). 

Contributed by c0sm0