EntityGraph: Bug fix: EntityGraphShape did not work correctly for EntityCollection properties in VB
Please see the following discussion for details.
http://riaservicescontrib.codeplex.com/discussions/264035

This patch will fix an issue that arises when using VB.Net. That 
language appears to wrap lambda expressions used in the definition 
of Edges in EntityGraphShapes which return EntityCollections in a 
Convert expression. The fix is to replace the EdgeEnumType overload 
of the EntityGraphShape.Edge method with the one in the attached file. 
The code checks for a UnaryExpression and extracts the operand.

Contributed by: samunro