EntityGraph: Added CopyTo extension method for EntityGraph shapes.
This method enables the construction of an entity graph of base type U
from an entity graph of base type T.

It can be used to e.g. import a graph of WCF RIA services entities from
a similarly named collection of entities from a non-RIA source. The other
way around is also possible.