EntityGraph: Improved Clone method; added Clone method that adds/attaches entities to a domain context
Clone method now uses Extract/ApplyState from EntityTools. This guarantees
that original/modified states are cloned as well.
    
There is a new Clone method that has a domain context as additional argument.
This method will automatically attach/add cloned entities to this context. This
guarantees that the EntityState of cloned entities is equal to the EntityState
of the original entities.